Change resource completion strategy
Change from a twofold resource completion strategy of completing the
resource provider and then the resource itself, to straight completion
of the resource itself.

Previously, the performance of this was very poor, but using
completionItem/resolve to fill in documentation on the fly (the bulk
of the content in any given CompletionItem) we can get away with it.

Now only the label and insert_text (for snippets) are sent by the
server on a textDocument/completion request on a resource.